requirements.txt
# Tool requirements.
crcmod
matplotlib
numpy
scipy
pyserial

# Testing requirements.
pytest
pytest-mock

# Formatting requirements.
black
